+++ thoth-app/src/component/about.rs
@@ -33,8 +33,8 @@ impl Component for AboutComponent {
                         <li>{ "Vincent W.J. van Gerven Oei (Director)" }</li>
                         <li>{ "Toby Steiner (COO)" }</li>
                         <li>{ "Javier Arias (CTO)" }</li>
-                        <li>{ "Brendan O'Connell (Software Engineer)" }</li>
                         <li>{ "Ross Higman (Software Engineer)" }</li>
+                        <li>{ "Brendan O'Connell (Software Engineer)" }</li>
                         <li>{ "Hannah Hillen (Metadata & Publisher Outreach Specialist)" }</li>
                     </ul>
                     <p class="title is-5">{ "Consultants" }</p>