+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[0.11.14]](https://github.com/thoth-pub/thoth/releases/tag/v0.11.14) - 2024-01-18
+### Changed
+  - [#467](https://github.com/thoth-pub/thoth/issues/467), [#403](https://github.com/thoth-pub/thoth/issues/403), [#536](https://github.com/thoth-pub/thoth/issues/536) - Expand the list of location platforms with: GoogleBooks, InternetArchive, ScienceOpen, and Scielo
+  - [526](https://github.com/thoth-pub/thoth/issues/526) - Added Brendan to About page
+
 ## [[0.11.13]](https://github.com/thoth-pub/thoth/releases/tag/v0.11.13) - 2024-01-08
 ### Changed
   - Upgrade rust to `1.75.0` in production and development `Dockerfile`